refactor(wizard): extract architecture prompts from bin/cli.js
Follows pattern from #81 to extract wizard logic into single-responsibility modules.

Changes:
- Created lib/wizard/arch-prompts.js (54 lines)
  - promptArchitectureGuardrails() with all prompt logic
  - Single responsibility: architecture configuration prompts

- Updated bin/cli.js
  - Replaced 30+ lines of inline prompts with helper call
  - Reduced from 764 → 734 lines (still needs more extraction)

Impact:
- Honors single responsibility principle
- Follows lib/wizard pattern from #81
- All tests passing (554)

Note: bin/cli.js still at 734 lines (target: ~100). Issue #76 refactor appears
incomplete - lib/commands/{init,learn,scaffold}.js files do not exist despite
PR #80 claiming to add them. Further refactoring needed to reach target.
